Filu/DB: Remove all sql files where only a DB function is called
In such a case to hardcode the sql in the C++ code should be Ok.
The idea behind the concept of holding the sql in a own file was
to have the possibility to change easy the sql without to recompile
the C++ code. But because we have already a own file for the sql
implementation (the DB function) looks it redundant.