Add delay at slot for watching used indicator for changes
I noticed a problem with rewrite of files: QFileSystemWatcher emits more
than one signal when a file is changed, with the result that the indicator
is reloaded twice. This delay did not fix this problem but give some time
that the file is intirely written on disk. Depending on the program which
made the change (Kate, nano, ManagerF) may still happen

  QFileSystemWatcher: failed to add paths: ...

nano and ManagerF did not cause this error.

See also: http://blog.rburchell.com/2012/03/qt-51-aka-when-qfilesystemwatcher-might.html